BackgroundDimmer rewrite in TS and removed its definition file.
