--- src/BackgroundDimmer/index.d.ts
@@ -1,3 +1,0 @@
-import BackgroundDimmer, { BackgroundDimmerProps } from './BackgroundDimmer';
-export default BackgroundDimmer;
-export { BackgroundDimmerProps };
