--- src/BackgroundDimmer/index.js
@@ -1,2 +1,0 @@
-import BackgroundDimmer from './BackgroundDimmer';
-export default BackgroundDimmer;
