+++ src/BackgroundDimmer/index.ts
@@ -1,0 +1,3 @@
+import BackgroundDimmer, { BackgroundDimmerProps, Props } from './BackgroundDimmer';
+export default BackgroundDimmer;
+export { BackgroundDimmerProps, Props };
